Implemented ChessServer\Socket\ChessSocket
