+++ src/Command/AbstractCommand.php
@@ -2,7 +2,7 @@
 
 namespace ChessServer\Command;
 
-use ChessServer\Socket;
+use ChessServer\Socket\ChessSocket;
 use Ratchet\ConnectionInterface;
 
 abstract class AbstractCommand
@@ -22,5 +22,5 @@ abstract class AbstractCommand
 
     abstract public function validate(array $command);
 
-    abstract public function run(Socket $socket, array $argv, ConnectionInterface $from);
+    abstract public function run(ChessSocket $socket, array $argv, ConnectionInterface $from);
 }
