+++ src/Command/AcceptPlayRequestCommand.php
@@ -3,7 +3,7 @@
 namespace ChessServer\Command;
 
 use Chess\Variant\Classical\PGN\AN\Color;
-use ChessServer\Socket;
+use ChessServer\Socket\ChessSocket;
 use ChessServer\Exception\InternalErrorException;
 use ChessServer\GameMode\PlayMode;
 use Ratchet\ConnectionInterface;
@@ -24,7 +24,7 @@ class AcceptPlayRequestCommand extends AbstractCommand
         return count($argv) - 1 === count($this->params);
     }
 
-    public function run(Socket $socket, array $argv, ConnectionInterface $from)
+    public function run(ChessSocket $socket, array $argv, ConnectionInterface $from)
     {
         $gameMode = $socket->getGameModeStorage()->getByHash($argv[1]);
