+++ src/Command/RematchCommand.php
@@ -2,7 +2,7 @@
 
 namespace ChessServer\Command;
 
-use ChessServer\Socket;
+use ChessServer\Socket\ChessSocket;
 use ChessServer\Exception\InternalErrorException;
 use ChessServer\GameMode\PlayMode;
 use Ratchet\ConnectionInterface;
@@ -38,7 +38,7 @@ class RematchCommand extends AbstractCommand
         return false;
     }
 
-    public function run(Socket $socket, array $argv, ConnectionInterface $from)
+    public function run(ChessSocket $socket, array $argv, ConnectionInterface $from)
     {
         $gameMode = $socket->getGameModeStorage()->getByResourceId($from->resourceId);
