+++ src/Command/StockfishEvalCommand.php
@@ -2,7 +2,7 @@
 
 namespace ChessServer\Command;
 
-use ChessServer\Socket;
+use ChessServer\Socket\ChessSocket;
 use ChessServer\Exception\InternalErrorException;
 use Ratchet\ConnectionInterface;
 
@@ -22,7 +22,7 @@ class StockfishEvalCommand extends AbstractCommand
         return count($argv) - 1 === count($this->params);
     }
 
-    public function run(Socket $socket, array $argv, ConnectionInterface $from)
+    public function run(ChessSocket $socket, array $argv, ConnectionInterface $from)
     {
         $gameMode = $socket->getGameModeStorage()->getByResourceId($from->resourceId);
